Added FragmentUtils
